make mlflow optional
